requirements.txt
google-genai
requests
pyyaml
moviepy
Pillow
python-dotenv  
streamlit
replicate
pydantic